Change $TERM to screen-256color
`rtv` was not working right and upon further investigation, found that
term type plays in important role.

Made this permanent, because mutt and scim are working better, while
better respecting bg colors. Also, the bg remains uniform and unlike
before, in scim, all empty cells are of the same color.