Learned about scrolloff in vim and `-e` in g-term
`scrolloff` keeps additional lines in the view port in vim during
scrolling. `L` and `H` are affected by this.

gnome-terminal can execute commands on start up using the `-e` option
and tmux can set the name of the session on start up using the `-s`
option. Combined and added them to bspwm.